Improvement on geom3d src, tests, and documentation
- Improved src code
- Improved documentation
- Organized tests
