+++ core/geometry/src/geometry3d/Point3dVector3d.ts
@@ -527,15 +527,20 @@ export class Point3d extends XYZ {
   public crossProductToPointsXY(pointA: Point3d, pointB: Point3d): number {
     return Geometry.crossProductXYXY(pointA.x - this.x, pointA.y - this.y, pointB.x - this.x, pointB.y - this.y);
   }
-  /** Return a point interpolated between this point and the right param. */
+  /**
+   * Return a point interpolated between `this` point and the `other` point.
+   * * fraction specifies where the interpolated point is located on the line passing `this` and `other`.
+   * */
   public interpolate(fraction: number, other: XYAndZ, result?: Point3d): Point3d {
     if (fraction <= 0.5)
-      return Point3d.create(this.x + fraction * (other.x - this.x),
+      return Point3d.create(
+        this.x + fraction * (other.x - this.x),
         this.y + fraction * (other.y - this.y),
         this.z + fraction * (other.z - this.z),
         result);
     const t: number = fraction - 1.0;
-    return Point3d.create(other.x + t * (other.x - this.x),
+    return Point3d.create(
+      other.x + t * (other.x - this.x),
       other.y + t * (other.y - this.y),
       other.z + t * (other.z - this.z),
       result);
@@ -902,8 +907,18 @@ export class Vector3d extends XYZ {
     this.z *= a;
     return true;
   }
-  /** Return the fractional projection of spaceVector onto this */
+  /**
+   * Return fractional projection of target vector onto this
+   * * It's returning the signed projection magnitude divided by the target magnitude. In other words,
+   * it's returning the length of the projection as a fraction of the target magnitude.
+   * @param target the target vector
+   * @param defaultFraction the returned value in case magnitude square of target vector is very small
+   * */
   public fractionOfProjectionToVector(target: Vector3d, defaultFraction: number = 0): number {
+    /*
+     * projection length is (this.target)/||target||
+     * but here we return (this.target)/||target||^2
+     */
     const numerator = this.dotProduct(target);
     const denominator = target.magnitudeSquared();
     if (denominator < Geometry.smallMetricDistanceSquared)
@@ -1161,7 +1176,8 @@ export class Vector3d extends XYZ {
     return true;
   }
   /**
-   * Compute cross product with `vectorB`.
+   * Compute cross product with `vectorB`
+   * * cross product vector will have the given length.
    * @param vectorB second vector for cross product.
    * @param productLength desired length of result vector.
    * @param result optional preallocated vector
@@ -1308,7 +1324,7 @@ export class Vector3d extends XYZ {
  * * The returned angle is always positive and no larger than 180 degrees (PI radians)
  * * The returned angle is "in the plane containing the two vectors"
  * * Use `planarRadiansTo` and `signedRadiansTo` to take have angle measured in specific plane.
- * @param vectorB target vector of rotation.
+ * @param vectorB target vector.
  */
   public radiansTo(vectorB: Vector3d): number {
     // ||axb|| = ||a|| ||b|| |sin(t)| and a.b = ||a|| ||b|| cos(t) ==>
@@ -1320,7 +1336,7 @@ export class Vector3d extends XYZ {
    * * The returned angle is always positive and no larger than 180 degrees (PI radians)
    * * The returned angle is "in the plane containing the two vectors"
    * * Use `planarAngleTo` and `signedAngleTo` to take have angle measured in specific plane.
-   * @param vectorB target vector of rotation.
+   * @param vectorB target vector.
    */
   public angleTo(vectorB: Vector3d): Angle {
     return Angle.createRadians(this.radiansTo(vectorB));
@@ -1340,7 +1356,7 @@ export class Vector3d extends XYZ {
    * * The returned angle can range from negative 180 degrees (negative PI radians) to positive 180
    * * degrees (positive PI radians), not closed on the negative side.
    * * Use `planarAngleTo` and `signedAngleTo` to take have angle measured in other planes.
-   * @param vectorB target vector of rotation.
+   * @param vectorB target vector.
    */
   public angleToXY(vectorB: Vector3d): Angle {
     return Angle.createAtan2(this.crossProductXY(vectorB), this.dotProductXY(vectorB));
@@ -1352,7 +1368,8 @@ export class Vector3d extends XYZ {
  * * The returned angle is "in the plane containing the two vectors"
  * * The returned angle has the same sign as vectorW dot product (thisVector cross vectorB)
  * * vectorW does not have to be perpendicular to the plane.
- * @param vectorB target vector of rotation.
+ * * Use planarRadiansTo to measure the angle between vectors that are projected to another plane.
+ * @param vectorB target vector.
  * @param vectorW distinguishes between the sides of the plane.
  */
   public signedRadiansTo(vectorB: Vector3d, vectorW: Vector3d): number {
@@ -1371,34 +1388,39 @@ export class Vector3d extends XYZ {
  * * The returned angle is "in the plane containing the two vectors"
  * * `vectorW` distinguishes between the sides of the plane, but does not have to be perpendicular.
  * * The returned angle has the same sign as vectorW dot product (thisVector cross vectorB)
- * @param vectorB target vector of rotation.
+ * * Use planarRadiansTo to measure the angle between vectors that are projected to another plane.
+ * @param vectorB target vector.
  * @param vectorW distinguishes between the sides of the plane.
  */
   public signedAngleTo(vectorB: Vector3d, vectorW: Vector3d): Angle {
     return Angle.createRadians(this.signedRadiansTo(vectorB, vectorW));
   }
   /**
-   * Return the (radians as a simple number, not strongly typed Angle) radians from this vector to vectorB.
+   * Return the radians (as a simple number, not strongly typed Angle) from this vector to vectorB,
+   * measured between their projections to the plane with the given normal.
    * * The returned angle can be positive or negative, with magnitude no larger than PI radians
-   * * Use signedRadiansTo` to take have angle measured in other planes.
-   * @param vectorB target vector of rotation.
-   * @param planeNormal a normal vector to the plane.
+   * @param vectorB target vector
+   * @param planeNormal the normal vector to the plane.
    */
   public planarRadiansTo(vectorB: Vector3d, planeNormal: Vector3d): number {
     const square = planeNormal.dotProduct(planeNormal);
     if (square === 0.0)
       return 0.0;
     const factor = 1.0 / square;
-    const projection0: Vector3d = this.plusScaled(planeNormal, -this.dotProduct(planeNormal) * factor);
-    const projection1: Vector3d = vectorB.plusScaled(planeNormal, -vectorB.dotProduct(planeNormal) * factor);
-    return projection0.signedRadiansTo(projection1, planeNormal);
+    /*
+     * projection of vector 'v' on normal 'n' is given by vProj = [dot(v,n)/||n||^2]*n
+     * and projection of 'v' on the plane is given by 'v - vProj'
+    */
+    const thisProj: Vector3d = this.plusScaled(planeNormal, -this.dotProduct(planeNormal) * factor);
+    const vectorBProj: Vector3d = vectorB.plusScaled(planeNormal, -vectorB.dotProduct(planeNormal) * factor);
+    return thisProj.signedRadiansTo(vectorBProj, planeNormal);
   }
   /**
-   * Return the (as strongly typed Angle) Angle from this vector to vectorB.
+   * Return the angle (as strongly typed Angle) from this vector to vectorB,
+   * measured between their projections to the plane with the given normal.
    * * The returned angle can range from negative PI to positive PI (not closed on negative side)
-   * * Use signedRadiansTo` to take have angle measured in other planes.
-   * @param vectorB target vector of rotation.
-   * @param planeNormal a normal vector to the plane.
+   * @param vectorB target vector.
+   * @param planeNormal the normal vector to the plane.
    */
   public planarAngleTo(vectorB: Vector3d, planeNormal: Vector3d): Angle {
     return Angle.createRadians(this.planarRadiansTo(vectorB, planeNormal));
