+++ core/geometry/src/geometry3d/Ray3d.ts
@@ -340,12 +340,12 @@ export class Ray3d implements BeJSONFunctions {
   }
 
   /**
-   * Return a ray whose ray.origin is interpolated, and ray.direction is the vector between points with a
-   * scale factor applied.
-   * @param pt1 start point of interpolation
+   * Return a ray whose `ray.origin` is an interpolated on the line passing pt1 and pt2 specified by the fraction
+   * and `ray.direction` is the vector from pt1 rto pt2 with a tangent scale factor applied.
+   * @param pt1 start point of interpolation.
    * @param fraction fractional position between points.
-   * @param pt2 endpoint of interpolation
-   * @param tangentScale scale factor to apply to the startToEnd vector
+   * @param pt2 endpoint of interpolation.
+   * @param tangentScale scale factor to apply to the startToEnd vector.
    * @param result optional receiver.
    */
   public static interpolatePointAndTangent(pt1: XYAndZ, fraction: number, pt2: XYAndZ, tangentScale: number, result?: Ray3d): Ray3d {
